refactor: Remove epub.ts
